# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个动画网站的后端API服务",
        "title": "动画网站 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/rss_feeds": {
            "get": {
                "description": "获取系统中所有已配置的RSS订阅源列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSS订阅源管理"
                ],
                "summary": "获取所有RSS订阅源",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.RSSResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.RSSFeedResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的RSS订阅源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSS订阅源管理"
                ],
                "summary": "创建RSS订阅源",
                "parameters": [
                    {
                        "description": "RSS订阅源信息",
                        "name": "feed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RSSFeedRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.RSSResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RSSFeedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    }
                }
            }
        },
...
        "/rss_feeds/{id}": {
            "get": {
                "description": "根据ID获取特定的RSS订阅源详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSS订阅源管理"
                ],
                "summary": "获取单个RSS订阅源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RSS订阅源ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.RSSResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RSSFeedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定ID的RSS订阅源信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSS订阅源管理"
                ],
                "summary": "更新RSS订阅源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RSS订阅源ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RSS订阅源信息",
                        "name": "feed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RSSFeedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.RSSResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RSSFeedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定ID的RSS订阅源",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSS订阅源管理"
                ],
                "summary": "删除RSS订阅源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RSS订阅源ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    }
                }
            }
        },
        "/rss_feeds/{id}/update": {
            "post": {
                "description": "手动触发指定ID的RSS订阅源的更新任务",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSS订阅源管理"
                ],
                "summary": "手动更新指定RSS订阅",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RSS订阅源ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.RSSResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.RSSResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.RSSFeedRequest": {
            "type": "object",
            "required": [
                "name",
                "parser_type",
                "update_interval",
                "url"
            ],
            "properties": {
                "keywords": {
                    "type": "string",
                    "example": "莉可丽丝,友谊是时间的窃贼"
                },
                "name": {
                    "type": "string",
                    "example": "莉可丽丝"
                },
                "page_end": {
                    "type": "integer",
                    "example": 5
                },
                "page_start": {
                    "type": "integer",
                    "example": 1
                },
                "parser_type": {
                    "type": "string",
                    "example": "mikanani"
                },
                "priority": {
                    "type": "integer",
                    "example": 0
                },
                "update_interval": {
                    "type": "integer",
                    "example": 1
                },
                "url": {
                    "type": "string",
                    "example": "https://mikanani.me/RSS/Bangumi?bangumiId=3644"
                }
            }
        },
        "models.RSSFeedResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keywords": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "page_end": {
                    "type": "integer"
                },
                "page_start": {
                    "type": "integer"
                },
                "parser_type": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "update_interval": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "请在此输入 Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}